+++ packages/expression_language/lib/src/visitors/expression_visitor.dart
@@ -42,6 +42,7 @@ abstract class ExpressionVisitor {
   void visitDelegate<T>(DelegateExpression<T> expression);
   void visitConditionalExpression<T>(ConditionalExpression<T> expression);
   void visitLengthFunction(LengthFunctionExpression expression);
+  void visitMatchesFunction(MatchesFunctionExpression expression);
   void visitToStringFunction(ToStringFunctionExpression expression);
   void visitIntToInteger(IntToIntegerExpression expression);
   void visitIntegerToInt(IntegerToIntExpression expression);
@@ -65,4 +66,7 @@ abstract class ExpressionVisitor {
       DurationInMinutesFunctionExpression expression);
   void visitDurationInSecondsFunction(
       DurationInSecondsFunctionExpression expression);
+  void visitContainsFunction(ContainsFunctionExpression expression);
+  void visitStartsWithFunction(StartsWithFunctionExpression expression);
+  void visitEndsWithFunction(EndsWithFunctionExpression expression);
 }