+++ packages/expression_language/CHANGELOG.md
@@ -1,10 +1,14 @@
+# 0.1.3
+
+Added matches, contains, startsWith and endsWith functions.
+
 # 0.1.2
 
-Fix small health suggestions.
+Fixed small health suggestions.
 
 # 0.1.1
 
-Fix formatting issues.
+Fixed formatting issues.
 
 # 0.1.0